Move get entry attributes to DataStore (#714)
* implement parse_onto_file

* implement parse_onto_file

* implement parse onto file v1; add _is_subclass signature

* fix typo

* add _get_entry_attributes_by_class to data store class

* move get_entry_attribute_by_class to data store

* fix pylint

* add OntologyCodeGenerator check to onto parse

* change black version to 22.3.0

* reformat with black

* remove unused import

* delete type_generator from data.rst

* update type_attribute and _parse_onto_file's docstring

* format

* delete parse onto file
