+++ forte/data/data_store.py
@@ -14,9 +14,9 @@
 from typing import List, Iterator, Tuple, Optional, Any
 import uuid
 from bisect import bisect_left
+
 from forte.utils import get_class
 from forte.data.base_store import BaseStore
-from forte.data.entry_type_generator import EntryTypeGenerator
 from forte.data.ontology.top import Annotation
 from forte.common import constants
 
@@ -122,37 +122,40 @@ class DataStore(BaseStore):
 
         """
         The ``_type_attributes`` is a private dictionary that provides
-        ``type_name`` and the order of corresponding attributes except ``index_id``.
-        The outer keys are indices of ``entry lists`` in ``self.__elements`` as
-        integers, representing all types that inherit the ``Entry`` class.
-        The inner keys are all the valid attributes for this type.
-        The values are the indices of attributes among these lists.
+        ``type_name``, their parent entry, and the order of corresponding attributes.
+        The keys are fully qualified names of every type; The value is a dictionary with
+        two keys. Key ``attribute`` provides an inner dictionary with all valid attributes
+        for this type and the indices of attributes among these lists. Key ``parent_entry``
+        is a string representing the direct parent of this type.
 
-        This structure is supposed to be obtained by processing the dictionary
-        generated by another function get_type_attributes(). This function will
-        be called in class ``EntryTypeGenerator``.
+        This structure is supposed to be built dynamically. When a user adds new entries,
+        data_store will check unknown types and add them to ``_type_attributes``.
 
         Example:
 
         .. code-block:: python
 
-            self._type_attributes = EntryTypeGenerator.get_type_attributes()
-
             # self._type_attributes is:
             # {
-            #     "ft.onto.base_ontology.Token": {"pos": 4, "ud_xpos": 5,
+            #     "ft.onto.base_ontology.Token": {
+            #       "attributes": {"pos": 4, "ud_xpos": 5,
             #               "lemma": 6, "chunk": 7, "ner": 8, "sense": 9,
             #               "is_root": 10, "ud_features": 11, "ud_misc": 12},
-            #     "ft.onto.base_ontology.Document": {"document_class": 4,
+            #       "parent_entry": "forte.data.ontology.top.Annotation", },
+            #     "ft.onto.base_ontology.Document": {
+            #       "attributes": {"document_class": 4,
             #               "sentiment": 5, "classifications": 6},
-            #     "ft.onto.base_ontology.Sentence": {"speaker": 4,
+            #       "parent_entry": "forte.data.ontology.top.Annotation", },
+            #     "ft.onto.base_ontology.Sentence": {
+            #       "attributes": {"speaker": 4,
             #               "part_id": 5, "sentiment": 6,
             #               "classification": 7, "classifications": 8},
+            #       "parent_entry": "forte.data.ontology.top.Annotation", }
             # }
         """
-        # Issue #570 implements get_type_attributes()
-        # see https://github.com/asyml/forte/issues/570
-        self._type_attributes: dict = EntryTypeGenerator.get_type_attributes()
+        self._type_attributes: dict = {}
+        if self.onto_file_path:
+            self._parse_onto_file()
 
         """
         The `__elements` is an underlying storage structure for all the entry
@@ -171,7 +174,7 @@ class DataStore(BaseStore):
         self.__elements: dict = {}
 
         """
-        A dictionary that keeps record of all entrys with their tid.
+        A dictionary that keeps record of all entries with their tid.
         It is a key-value map of {tid: entry data in list format}.
 
         e.g., {1423543453: [begin, end, tid, type_name, attr_1, ..., attr_n],
@@ -546,3 +549,66 @@ class DataStore(BaseStore):
             The previous entry of the same type as the ``tid`` entry.
         """
         raise NotImplementedError
+
+    def _parse_onto_file(self):
+        r"""This function will populate the types and attributes used in the data_store
+        with an ontology specification file. If a user provides a customized ontology
+        specification file, forte will parse this file and set the internal dictionary
+        ``self._type_attributes`` to store type name, parent entry, and its attribute
+        information accordingly.
+
+        For every ontology, this function will import paths containing its parent entry and
+        merge all classes contained in the imported file into the dictionary. For example,
+        if an ontology has a parent entry in ``ft.onto.base_ontology``, all classes in
+        ``ft.onto.base_ontology`` will be imported and stored in the internal dictionary.
+        A user can use classes both in the ontology specification file and their parent
+        entries's paths.
+        """
+        if self.onto_file_path is None:
+            return
+        raise NotImplementedError
+
+    def _get_entry_attributes_by_class(
+        self, input_entry_class_name: str
+    ) -> List:
+        """Get type attributes by class name. `input_entry_class_name` should be
+        a fully qualified name of an entry class.
+
+        The `dataclass` module<https://docs.python.org/3/library/dataclasses.html> can add
+        generated special methods to user-defined classes. There is an in-built function
+        called `__dataclass_fields__` that is called on the class object, and it returns
+        all the fields the class contains.
+
+        .. note::
+
+            This function is only applicable to classes decorated as Python
+            `dataclass` since it relies on the `__dataclass_fields__` to find out the attributes.
+
+
+        Args:
+            input_entry_class_name: A fully qualified name of an entry class.
+
+        Returns:
+            A list of attributes corresponding to the input class.
+
+        For example, for Sentence we want to get a list of
+        ["speaker", "part_id", "sentiment", "classification", "classifications"].
+        The solution looks like the following:
+
+        .. code-block:: python
+
+            # input can be a string
+            entry_name = "ft.onto.base_ontology.Sentence"
+
+            # function signature
+            get_entry_attributes_by_class(entry_name)
+
+            # return
+            # ["speaker", "part_id", "sentiment", "classification", "classifications"]
+
+        """
+        class_ = get_class(input_entry_class_name)
+        try:
+            return list(class_.__dataclass_fields__.keys())
+        except AttributeError:
+            return []
